LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/register-same-scope-different-script-url.https-expected.txt



FAIL Register different scripts concurrently assert_unreached: unexpected rejection: assert_equals: on second register, second script should be installing expected "https://localhost:9443/service-workers/service-worker/resources/empty-worker.js?new" but got "https://localhost:9443/service-workers/service-worker/resources/empty-worker.js" Reached unreachable code
FAIL Register then register new script URL assert_unreached: unexpected rejection: assert_equals: on second register, the second script should be installing expected "https://localhost:9443/service-workers/service-worker/resources/empty-worker.js?new" but got "https://localhost:9443/service-workers/service-worker/resources/empty-worker.js" Reached unreachable code
FAIL Register then register new script URL that 404s assert_unreached: unexpected rejection: assert_unreached: register should reject Reached unreachable code Reached unreachable code
FAIL Register then register new script that does not install assert_unreached: unexpected rejection: assert_equals: on update, the second script should be installing expected "https://localhost:9443/service-workers/service-worker/resources/reject-install-worker.js" but got "https://localhost:9443/service-workers/service-worker/resources/empty-worker.js" Reached unreachable code
FAIL Register same-scope new script url effect on controller assert_unreached: unexpected rejection: null is not an object (evaluating 'sw_container.controller.scriptURL') Reached unreachable code
